test(browser): extract shared browser-suite helpers into tests/Browser/helpers.php
The browser suite runs against the assembled core app, where Pest only auto-loads
the ROOT tests/Pest.php — a Pest.php synced into tests/Browser/web is never loaded.
So each test file had re-declared the same helpers (deviceVisit, snap,
actingAsCharacter, realCharacterId, the seeders …) under function_exists guards.

Move every helper into one require_once'd tests/Browser/helpers.php (idempotent,
so a single definition is safe across a multi-file run) and slim each test file to
`require_once __DIR__.'/helpers.php';`. actingAsCharacter/giveCorporationRole keep
their function_exists guard so they coexist with a core Pest.php that still
declares them during the transition. Pure move — no behavioural change.
