+++ tests/Browser/RecruitmentLifecycleTest.php
@@ -1,26 +1,13 @@
 <?php
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
-use Pest\Browser\Api\PendingAwaitablePage;
-use Pest\Browser\Enums\Device;
-use Pest\Browser\Playwright\Playwright;
-use Seatplus\Auth\Models\AccessControl\RoleMembership;
-use Seatplus\Auth\Models\CharacterUser;
-use Seatplus\Auth\Models\Permissions\Permission;
-use Seatplus\Auth\Models\Permissions\Role;
 use Seatplus\Auth\Models\User;
 use Seatplus\Eveapi\Models\Application;
 use Seatplus\Eveapi\Models\Character\CharacterInfo;
-use Seatplus\Eveapi\Models\Character\CharacterRole;
 use Seatplus\Eveapi\Models\Recruitment\ApplicationLogs;
-use Seatplus\Eveapi\Models\Skills\Skill;
-use Seatplus\Eveapi\Models\Skills\SkillQueue;
-use Seatplus\Eveapi\Models\Universe\Category;
-use Seatplus\Eveapi\Models\Universe\Group;
-use Seatplus\Eveapi\Models\Universe\Type;
 use Seatplus\Eveapi\Models\Wallet\WalletJournal;
-use Seatplus\Web\Models\Recruitment\Enlistment;
-use Seatplus\Web\Models\Recruitment\EnlistmentReviewRound;
+
+require_once __DIR__.'/helpers.php';
 
 /*
  * End-to-end browser tests for the employee-recruitment lifecycle, run against the assembled core
@@ -33,185 +20,6 @@ use Seatplus\Web\Models\Recruitment\EnlistmentReviewRound;
  */
 
 uses(RefreshDatabase::class);
-
-if (! function_exists('deviceVisit')) {
-    function deviceVisit(string $device, string $url, array $options = []): mixed
-    {
-        if ($device === 'iphone') {
-            return new PendingAwaitablePage(Playwright::defaultBrowserType(), Device::IPHONE_15, $url, $options);
-        }
-
-        return visit($url, $options);
-    }
-}
-
-if (! function_exists('snap')) {
-    function snap($page, string $name): void
-    {
-        $page->script("document.querySelectorAll('img').forEach((i) => { i.loading = 'eager'; });");
-        $page->waitForEvent('networkidle');
-        $page->screenshot(true, $name);
-    }
-}
-
-if (! function_exists('userOfCharacter')) {
-    function userOfCharacter(int $characterId): User
-    {
-        return CharacterUser::query()->where('character_id', $characterId)->firstOrFail()->user;
-    }
-}
-
-if (! function_exists('realCharacterId')) {
-    // images.evetech.net serves a generic default for fabricated ids, so seeded applicants must use a
-    // real EVE character id to render a real portrait in the screenshots. Shared pool across the suite.
-    function realCharacterId(): int
-    {
-        $pool = [197343093, 1319140135, 92081232, 1191750472, 94391213, 887625289, 1435633555, 1809892636];
-        $available = array_values(array_diff($pool, CharacterInfo::query()->pluck('character_id')->all()));
-
-        return $available[0] ?? fake()->unique()->numberBetween(9000000, 98000000);
-    }
-}
-
-if (! function_exists('makeControlGroup')) {
-    function makeControlGroup(string $name): Role
-    {
-        return Role::findById(Role::create(['name' => $name])->id);
-    }
-}
-
-if (! function_exists('addGroupMember')) {
-    function addGroupMember(Role $role, User $user): void
-    {
-        RoleMembership::create([
-            'role_id' => $role->id,
-            'entity_type' => User::class,
-            'entity_id' => $user->getKey(),
-            'status' => 'active',
-        ]);
-        $user->assignRole($role);
-    }
-}
-
-if (! function_exists('makeRecruiterOfCorporation')) {
-    /** Grant the acting user recruiter access to their corporation (permission + director role). */
-    function makeRecruiterOfCorporation(CharacterInfo $character, string $permission): User
-    {
-        $user = userOfCharacter($character->character_id);
-        $user->givePermissionTo(Permission::findOrCreate($permission));
-
-        CharacterRole::updateOrCreate(
-            ['character_id' => $character->character_id],
-            ['roles' => ['Director']],
-        );
-
-        return $user;
-    }
-}
-
-if (! function_exists('openPostingWithStages')) {
-    /**
-     * @param  array<int, array{label: string, role: ?Role}>  $stages
-     */
-    function openPostingWithStages(int $corporationId, array $stages): void
-    {
-        Enlistment::query()->updateOrCreate(['corporation_id' => $corporationId], ['type' => 'user']);
-
-        foreach ($stages as $position => $stage) {
-            EnlistmentReviewRound::query()->updateOrCreate(
-                ['corporation_id' => $corporationId, 'position' => $position],
-                ['label' => $stage['label'], 'role_id' => $stage['role']?->id],
-            );
-        }
-    }
-}
-
-if (! function_exists('seedApplicantAtStage')) {
-    /** Create an applicant with an open application to $corporationId, already advanced to $stage. */
-    function seedApplicantAtStage(int $corporationId, string $name, int $stage = 0): Application
-    {
-        $character = CharacterInfo::factory()->create(['character_id' => realCharacterId(), 'name' => $name]);
-
-        $user = new User;
-        $user->main_character_id = $character->character_id;
-        $user->save();
-
-        CharacterUser::create([
-            'user_id' => $user->getKey(),
-            'character_id' => $character->character_id,
-            'character_owner_hash' => sha1((string) $character->character_id),
-        ]);
-
-        $application = Application::query()->create([
-            'corporation_id' => $corporationId,
-            'applicationable_type' => User::class,
-            'applicationable_id' => $user->getKey(),
-        ]);
-
-        // Each recorded decision advances the application one stage.
-        for ($i = 0; $i < $stage; $i++) {
-            ApplicationLogs::query()->create([
-                'application_id' => $application->id,
-                'causer_type' => User::class,
-                'causer_id' => $user->getKey(),
-                'type' => 'decision',
-                'comment' => '',
-            ]);
-        }
-
-        return $application;
-    }
-}
-
-if (! function_exists('makeSkillType')) {
-    /**
-     * @param  array{type: int, type_name: string, group: int, group_name: string}  $skill
-     */
-    function makeSkillType(array $skill): Type
-    {
-        if (! Category::query()->whereKey(16)->exists()) {
-            Category::factory()->create(['category_id' => 16, 'name' => 'Skill', 'published' => true]);
-        }
-        if (! Group::query()->whereKey($skill['group'])->exists()) {
-            Group::factory()->create(['group_id' => $skill['group'], 'category_id' => 16, 'name' => $skill['group_name'], 'published' => true]);
-        }
-
-        return Type::query()->whereKey($skill['type'])->first()
-            ?? Type::factory()->create(['type_id' => $skill['type'], 'group_id' => $skill['group'], 'name' => $skill['type_name'], 'published' => true]);
-    }
-}
-
-if (! function_exists('seedCharacterSkills')) {
-    /** Seed a small deterministic skill sheet for $character; returns the queued skill's name. */
-    function seedCharacterSkills(CharacterInfo $character): string
-    {
-        $gunnery = makeSkillType(['type' => 3300, 'type_name' => 'Gunnery', 'group' => 255, 'group_name' => 'Gunnery']);
-        $smallHybrid = makeSkillType(['type' => 3301, 'type_name' => 'Small Hybrid Turret', 'group' => 255, 'group_name' => 'Gunnery']);
-        $spaceshipCommand = makeSkillType(['type' => 3327, 'type_name' => 'Spaceship Command', 'group' => 257, 'group_name' => 'Spaceship Command']);
-
-        foreach ([$gunnery, $smallHybrid, $spaceshipCommand] as $type) {
-            Skill::factory()->create([
-                'character_id' => $character->character_id,
-                'skill_id' => $type->type_id,
-                'active_skill_level' => 5,
-                'trained_skill_level' => 5,
-                'skillpoints_in_skill' => 256_000,
-            ]);
-        }
-
-        $queued = makeSkillType(['type' => 3302, 'type_name' => 'Small Projectile Turret', 'group' => 255, 'group_name' => 'Gunnery']);
-        SkillQueue::factory()->create([
-            'character_id' => $character->character_id,
-            'skill_id' => $queued->type_id,
-            'queue_position' => 0,
-            'finished_level' => 5,
-            'start_date' => now()->subDay()->format('Y-m-d H:i:s'),
-            'finish_date' => now()->addDays(3)->format('Y-m-d H:i:s'),
-        ]);
-
-        return $queued->name;
-    }
-}
 
 // ═══ HR manager — Manage Recruitment ══════════════════════════════════════════════════════════
